Rename 'homewizard_energy' to 'homewizard'
