+++ homeassistant/components/homewizard/coordinator.py
