+++ homeassistant/components/homewizard/sensor.py
