+++ homeassistant/components/homewizard/strings.json
