+++ tests/components/homewizard/generator.py
